Use separate routing contexts for anonymous and normal routes
