+++ mathesar_ui/src/components/routing/RouteObserver.svelte
@@ -8,17 +8,25 @@
     unload: undefined;
   }>();
 
-  export let meta: TinroRouteMeta;
+  export let meta: TinroRouteMeta | undefined = undefined;
 
   onMount(() => {
-    const unsubsriber = $meta.subscribe((metaInfo) => {
-      if (metaInfo) {
-        dispatch('load', metaInfo);
-      }
-    });
+    if (meta?.subscribe) {
+      const unsubsriber = meta.subscribe((metaInfo) => {
+        if (metaInfo) {
+          dispatch('load', metaInfo);
+        }
+      });
+
+      return () => {
+        unsubsriber();
+        dispatch('unload');
+      };
+    }
+
+    dispatch('load');
 
     return () => {
-      unsubsriber();
       dispatch('unload');
     };
   });
