+++ mathesar_ui/src/routes/AnonymousAccessRoutes.svelte
